Sample.video.mp4
This is a video that showcases the final output of the project. It seems to be able to predict some emotions quite accurately. This is not the clearest video as there was some problem with my system, so I made it from my smartphone but still, it showcases the output pretty well.